+++ frontend/src/App.tsx
@@ -2,8 +2,8 @@ import React from "react";
 
 import { LayoutElement, Workbench } from "@framework/Workbench";
 import { Content } from "@framework/components/Content";
-import { Settings } from "@framework/components/Settings";
-import { TopNavBar } from "@framework/components/TopNavBar";
+import { Settings } from "@framework/internal/components/Settings";
+import { TopNavBar } from "@framework/internal/components/TopNavBar";
 import { AuthProvider } from "@framework/providers/AuthProvider";
 import { CustomQueryClientProvider } from "@framework/providers/QueryClientProvider";
 import { ResizablePanels } from "@lib/components/ResizablePanels";
