+++ frontend/src/framework/internal/components/Content/content.tsx
