+++ frontend/src/framework/internal/components/Content/index.ts
