+++ frontend/src/framework/internal/components/Content/private-components/ViewWrapper/index.ts
