+++ frontend/src/framework/internal/components/Content/private-components/ViewWrapper/private-components/header.tsx
