+++ frontend/src/framework/internal/components/Content/private-components/ViewWrapper/viewWrapper.tsx
@@ -2,12 +2,11 @@ import React from "react";
 
 import { ModuleInstance } from "@framework/ModuleInstance";
 import { Workbench } from "@framework/Workbench";
-import { Point, pointRelativeToDomRect, pointerEventToPoint } from "@framework/utils/geometry";
+import { Point, pointRelativeToDomRect, pointerEventToPoint, pointDifference } from "@framework/utils/geometry";
 
 import { Header } from "./private-components/header";
 import { ViewContent } from "./private-components/viewContent";
 
-import { pointDifference } from "../../../../utils/geometry";
 import { LayoutEventTypes } from "../layout";
 import { ViewWrapperPlaceholder } from "../viewWrapperPlaceholder";
