+++ frontend/src/framework/internal/components/Content/private-components/drawer.tsx
