+++ frontend/src/framework/internal/components/Content/private-components/layoutBox.tsx
