+++ frontend/src/framework/internal/components/Content/private-components/syncSettings.tsx
