+++ frontend/src/framework/internal/components/Content/private-components/viewWrapperPlaceholder.tsx
