+++ frontend/src/framework/internal/components/ErrorBoundary/index.ts
